data-hub/requirements.txt
# setuptools==69.0.3
# six==1.16.0
# wheel==0.42.0
# requests==2.31.0
# gdal==2.4.0
#urllib3==1.26.5
#shapely==2.0.2
#pillow==10.2.0


Django==5.1.4
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.6.0
django-uw-keycloak==2.0.2
#python-keycloak==5.1.1
python-jose==3.3.0
python-dotenv==1.0.1

#psycopg2-binary==2.9.9
psycopg2==2.9.10 

gunicorn==23.0.0 
uvicorn[standard]


celery==5.4.0 
redis==5.2.1

channels[daphne]==4.2.0
channels-redis==4.2.1 

numpy==2.2.0 

pandas